% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vad.R
\name{is.voiced}
\alias{is.voiced}
\title{Get from a Voice Activity Detection (VAD object) the segments which are voiced}
\usage{
is.voiced(x, ...)
}
\arguments{
\item{x}{an object of class VAD}

\item{...}{further arguments passed on to the function}

\item{units}{character string with the units to use - either 'seconds' or 'milliseconds'}
}
\value{
A data.frame with columns vad_segment, start, end, duration, has_voice indicating where in the audio voice is detected
}
\description{
Postprocessing the Voice Activity Detection whereby sequences of 
voiced/non-voiced segments are collapsed by 
\enumerate{
\item{first considering all non-voiced segments which are small in duration (default < 1 second) voiced}
\item{next considering voiced segments with length less than a number of seconds (default < 1 second) non-voiced}
}
}
\examples{
file   <- system.file(package = "audio.vadwebrtc", "extdata", "test_wav.wav")
vad    <- VAD(file, mode = "normal", milliseconds = 30)
vad$vad_segments
voiced <- is.voiced(vad, silence_min = 0.2)
voiced
voiced <- is.voiced(vad, silence_min = 200, units = "milliseconds")
voiced
}
