% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vad.R
\name{VAD_channel}
\alias{VAD_channel}
\title{Voice Activity Detection per channel}
\usage{
VAD_channel(file, sample_rate = 16000, channels = c("default", "all"), ...)
}
\arguments{
\item{file}{the path to an audio file}

\item{sample_rate}{integer with the \code{sample_rate} to convert the file to. Passed on to \code{\link[av]{av_audio_convert}}}

\item{channels}{character string - either 'default' or 'all' indicating to do the voice activity detection for each channel independently ('default') or for all channels independently as well as all channels together ('all')}

\item{...}{further arguments passed on to \code{\link{VAD}}}
}
\value{
an object of class \code{webrtc-gmm-bychannel} which is a list with elements
\itemize{
\item{file: the path to the file}
\item{duration_secs: seconds}
\item{sample_rate: the sample rate of the audio file in Hz}
\item{channels: the number of channels in the audio}
\item{samples: the number of samples in the data}
\item{bitsPerSample: the number of bits per sample}
\item{bytesPerSample: the number of bytes per sample}
\item{type: the type of VAD model - currently only 'webrtc-gmm'}
\item{mode: the provided VAD mode}
\item{milliseconds: the provided milliseconds - either by 10, 20 or 30 ms frames}
\item{frame_length: the frame length corresponding to the provided milliseconds}
\item{vad_segments: a data.frame with columns channel, vad_segment, start, end and has_voice where the start/end values are in seconds}
\item{vad_stats: a list with elements channel, n_segments, n_segments_has_voice, n_segments_has_no_voice, seconds_has_voice, seconds_has_no_voice, pct_has_voice indicating the number of segments with voice and the duration of the voice/non-voice in the audio}
}
Channel 0 means all audio combined in 1 channel.
}
\description{
Voice Activity Detection per channel. 
Transforms the audio file to a wav file with the provided \code{sample_rate} and perform the voice activity detection per channel.
}
\examples{
library(audio)
library(av)
file <- system.file(package = "audio.vadwebrtc", "extdata", "stereo.mp3")
vad  <- VAD_channel(file, sample_rate = 32000, 
                    mode = "normal", milliseconds = 10, channels = "all")
vad
vad$vad_segments
voiced <- is.voiced(vad, channel = 0, silence_min = 0.2, voiced_min = 1)
voiced
voiced <- is.voiced(vad, channel = 1, silence_min = 0.2, voiced_min = 1)
voiced
voiced <- is.voiced(vad, channel = 2, silence_min = 0.2, voiced_min = 1)
voiced
}
