% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vad.R
\name{VAD}
\alias{VAD}
\title{Voice Activity Detection}
\usage{
VAD(
  file,
  mode = c("normal", "lowbitrate", "aggressive", "veryaggressive"),
  type = "webrtc"
)
}
\arguments{
\item{file}{the path to an audio file which should be a file in 16 bit with mono PCM samples (pcm_s16le codec) with a sampling rate of either 8Khz, 16KHz or 32Khz}

\item{mode}{the type of voice detection, either 'normal', 'lowbitrate', 'aggressive' or 'veryaggressive'}

\item{type}{character string with the type of VAD model. Only 'webrtc' currently.}
}
\value{
a list with elements
\itemize{
\item{file}
\item{sample_rate}
\item{channels}
\item{samples}
\item{format}
\item{bitsPerSample}
\item{bytesPerSample}
\item{vad: a data.frame with columns millisecond, has_voice and vad_segment indicating if the audio contains an active voice signal at that millisecond}
\item{vad_segments: a data.frame with columns vad_segment, start, end and has_voice where the start/end values are in seconds}
}
}
\description{
Detect the location of active Voice Activity Detection using webrtc
}
\examples{
file <- system.file(package = "audio.vadwebrtc", "extdata", "test_wav.wav")
vad  <- VAD(file, mode = "normal")
vad$vad_segments
\dontrun{
library(av)
x <- read_audio_bin(file)
plot(seq_along(x) / 16000, x, type = "l")
abline(v = vad$vad_segments$start, col = "red", lwd = 2)
abline(v = vad$vad_segments$end, col = "blue", lwd = 2)

av_media_info(file)
av_audio_convert(file, output = "audio_pcm_16khz.wav", 
                 format = "wav", channels = 1, sample_rate = 16000)
vad <- VAD("audio_pcm_16khz.wav", mode = "normal")
}
}
